+++ packages/webcomponents/src/api/TerminalOrder.ts
@@ -1,12 +1,20 @@
 import { TerminalProviders } from './Terminal';
 import { TerminalModelName } from './TerminalModel';
 
-enum TerminalOrderType {
+export interface TerminalOrderQueryParams {
+  order_status?: TerminalOrderStatus;
+  order_type?: TerminalOrderType;
+  created_after?: string;
+  created_before?: string;
+}
+
+export enum TerminalOrderType {
   boardingOnly = 'boarding_only',
   boardingShipping = 'boarding_shipping',
 }
 
 export enum TerminalOrderStatus {
+  created = 'created',
   submitted = 'submitted',
   completed = 'completed',
 }
@@ -27,7 +35,7 @@ interface TerminalOrderItem {
 export interface ITerminalOrder {
   id?: string;
   business_id?: string;
-  account_id?: string;
+  sub_account_id?: string;
   provider?: TerminalProviders;
   order_type?: TerminalOrderType;
   order_status?: TerminalOrderStatus;
@@ -39,7 +47,7 @@ export interface ITerminalOrder {
 export class TerminalOrder {
   public id: string;
   public business_id: string;
-  public account_id: string;
+  public sub_account_id: string;
   public provider: TerminalProviders;
   public order_type: TerminalOrderType;
   public order_status: TerminalOrderStatus;
@@ -51,7 +59,7 @@ export class TerminalOrder {
   constructor(data: ITerminalOrder) {
     this.id = data.id;
     this.business_id = data.business_id;
-    this.account_id = data.account_id;
+    this.sub_account_id = data.sub_account_id;
     this.provider = data.provider || TerminalProviders.verifone;
     this.order_items = [];
     this.order_type = data.order_type;
@@ -64,8 +72,7 @@ export class TerminalOrder {
   public get payload() {
     return {
       business_id: this.business_id,
-      account_id: this.account_id,
-      provider: this.provider,
+      sub_account_id: this.sub_account_id,
       order_type: this.order_type,
       order_items: this.order_items,
     };