+++ packages/webcomponents/src/components/terminal-orders-list/terminal-orders-list.tsx
@@ -1,7 +1,11 @@
-import { Component, h, Prop, State } from '@stencil/core';
+import { Component, h, Prop, State, Event, EventEmitter } from '@stencil/core';
 import { StyledHost } from '../../ui-components';
 import { checkPkgVersion } from '../../utils/check-pkg-version';
 import { defaultColumnsKeys } from './terminal-orders-table';
+import { makeGetTerminalOrders } from '../../actions/terminal/get-terminal-orders';
+import { TerminalOrderService } from '../../api/services/terminal_orders.service';
+import { ComponentErrorCodes, ComponentErrorEvent, ComponentErrorSeverity } from '../../api';
+import JustifiAnalytics from '../../api/Analytics';
 
 @Component({
   tag: 'justifi-terminal-orders-list',
@@ -12,31 +16,50 @@ export class TerminalOrdersList {
   @State() getTerminalOrders: Function;
   @State() errorMessage: string = null;
 
-  @Prop() businessId: string;
+  @Prop() accountId: string;
   @Prop() authToken: string;
   @Prop() apiOrigin?: string = PROXY_API_ORIGIN;
   @Prop() columns?: string = defaultColumnsKeys
-  
+
+  @Event({ eventName: 'error-event' }) errorEvent: EventEmitter<ComponentErrorEvent>;
+
+  analytics: JustifiAnalytics;
+
   componentWillLoad() {
     checkPkgVersion();
-    this.initializeGetTerminalOrders();
+    this.analytics = new JustifiAnalytics(this);
+    this.initializeGetData();
   }
 
-  private initializeGetTerminalOrders() {
-    if (this.businessId && this.authToken) {
-      this.getTerminalOrders = () => {
-        console.log('getTerminalOrders');
-      }
+  private initializeGetData() {
+    if (this.accountId && this.authToken) {
+      this.getTerminalOrders = makeGetTerminalOrders({
+        id: this.accountId,
+        authToken: this.authToken,
+        service: new TerminalOrderService(),
+        apiOrigin: this.apiOrigin
+      });
     } else {
-      this.errorMessage = 'Business ID and Auth Token are required';
+      this.errorMessage = 'Account ID and Auth Token are required';
+      this.errorEvent.emit({
+        errorCode: ComponentErrorCodes.MISSING_PROPS,
+        message: this.errorMessage,
+        severity: ComponentErrorSeverity.ERROR,
+      });
     }
   }
+
+  handleErrorEvent = (event) => {
+    this.errorMessage = event.detail.message;
+    this.errorEvent.emit(event.detail);
+  };
 
   render() {
     return (
       <StyledHost>
-        <terminal-orders-list-core 
-          getTerminalOrders={this.getTerminalOrders} 
+        <terminal-orders-list-core
+          getTerminalOrders={this.getTerminalOrders}
+          onError-event={this.handleErrorEvent}
           columns={this.columns}
         />
       </StyledHost>