+++ packages/webcomponents/src/components/terminal-orders-list/test/terminal-orders-list-with-filters.spec.tsx
@@ -1,0 +1,145 @@
+jest.mock('../../../ui-components/styled-host/styled-host.css', () => '');
+
+import { h } from '@stencil/core';
+import { newSpecPage } from '@stencil/core/testing';
+import { TerminalOrdersList } from '../terminal-orders-list';
+import { defaultColumnsKeys } from '../terminal-orders-table';
+import { TableFiltersMenu } from '../../../ui-components/filters/table-filters-menu';
+import { TerminalOrdersListFilters } from '../terminal-orders-list-filters';
+import { SelectInput } from '../../../ui-components/form/form-control-select';
+import { filterParams } from '../terminal-orders-list-params-state';
+import { TerminalOrdersListCore } from '../terminal-orders-list-core';
+
+const components = [TerminalOrdersList, TerminalOrdersListCore, TableFiltersMenu, TerminalOrdersListFilters, SelectInput];
+
+describe('justifi-terminal-orders-list with filters', () => {
+  const fetchDataSpy = jest.spyOn(TerminalOrdersListCore.prototype, 'fetchData');
+  let page;
+
+  afterEach(() => {
+    page = null;
+    jest.resetAllMocks();
+  });
+
+  it('shows table filter menu on filter button click', async () => {
+
+    page = await newSpecPage({
+      components: components,
+      autoApplyChanges: true,
+      flushQueue: true,
+      template: () => 
+      <div>
+        <justifi-terminal-orders-list-filters />
+        <justifi-terminal-orders-list 
+          account-id="abc"
+          auth-token="abc" 
+          columns={defaultColumnsKeys} 
+        />,
+      </div>
+    });
+
+    const filterButton = page.root.shadowRoot.querySelector('[data-test-id="open-filters-button"]') as HTMLElement;
+    expect(filterButton).not.toBeNull();
+
+    filterButton.click();
+
+    const filterMenu = page.root.shadowRoot.querySelector('[data-test-id="filter-menu"]') as HTMLElement;
+    expect(filterMenu).not.toBeNull();
+  });
+
+  it('updates params and refetches data on filter interaction', async () => {
+
+    page = await newSpecPage({
+      components: components,
+      autoApplyChanges: true,
+      template: () => 
+      <div>
+        <justifi-terminal-orders-list-filters />
+        <justifi-terminal-orders-list 
+          account-id="abc"
+          auth-token="abc" 
+          columns={defaultColumnsKeys}
+        />,
+      </div>
+    });
+
+    expect(fetchDataSpy).toHaveBeenCalled();
+
+    const filterButton = page.root.shadowRoot.querySelector('[data-test-id="open-filters-button"]') as HTMLElement;
+    filterButton.click();
+
+    const filterMenu = page.root.shadowRoot.querySelector('[data-test-id="filter-menu"]') as HTMLElement;
+    expect(filterMenu).not.toBeNull();
+
+    const selectFilter = filterMenu.querySelector('form-control-select') as HTMLFormControlSelectElement;
+    expect(selectFilter).not.toBeNull();
+
+    const selectFilterInput = selectFilter.querySelector('select') as HTMLSelectElement;
+    expect(selectFilterInput).not.toBeNull();
+
+    selectFilterInput.click();
+    
+    const selectOptions = selectFilterInput.querySelectorAll('option');
+    expect(selectOptions).not.toBeNull();
+    
+    const succeededOption = selectOptions[3] as HTMLOptionElement;
+    succeededOption.click();
+    selectFilterInput.value = 'completed';
+    selectFilterInput.dispatchEvent(new Event('input'));
+
+    expect(fetchDataSpy).toHaveBeenCalled();
+    const updatedParams = filterParams;
+    expect(updatedParams).toEqual({"order_status": "completed"});
+  });
+
+  it('clears filters and refetches data on clear filters interaction', async () => {
+
+    page = await newSpecPage({
+      components: components,
+      autoApplyChanges: true,
+      flushQueue: true,
+      template: () => 
+      <div>
+        <justifi-terminal-orders-list-filters />
+        <justifi-terminal-orders-list 
+          account-id="abc"
+          auth-token="abc" 
+          columns={defaultColumnsKeys} 
+        />,
+      </div>
+    });
+
+    expect(fetchDataSpy).toHaveBeenCalled();
+
+    const filterButton = page.root.shadowRoot.querySelector('[data-test-id="open-filters-button"]') as HTMLElement;
+    filterButton.click();
+
+    const filterMenu = page.root.shadowRoot.querySelector('[data-test-id="filter-menu"]') as HTMLElement;
+    expect(filterMenu).not.toBeNull();
+
+    const selectFilter = page.root.shadowRoot.querySelector('form-control-select') as HTMLFormControlSelectElement;
+    expect(selectFilter).not.toBeNull();
+
+    const selectFilterInput = selectFilter.querySelector('select') as HTMLSelectElement;
+    expect(selectFilterInput).not.toBeNull();
+
+    selectFilterInput.click();
+    
+    const selectOptions = selectFilterInput.querySelectorAll('option');
+    expect(selectOptions).not.toBeNull();
+    
+    const succeededOption = selectOptions[3] as HTMLOptionElement;
+    succeededOption.click();
+    selectFilterInput.value = 'succeeded';
+    selectFilterInput.dispatchEvent(new Event('input'));
+
+    const clearButton = page.root.shadowRoot.querySelector('[data-test-id="clear-filters-button"]') as HTMLElement;
+    expect(clearButton).not.toBeNull();
+
+    clearButton.click();
+
+    expect(fetchDataSpy).toHaveBeenCalled();
+    const updatedParams = page.rootInstance.params;
+    expect(updatedParams).toEqual(undefined);
+  });
+});